[VideoInfoScanner] Move the artwork side of a scrape into its own class.
